Add  authentication

+++ src/Encryption/SymmetricEncryption.cs
@@ -10,7 +10,7 @@ namespace Encryption
     {
         public static void Decrypt(Stream input, Stream output, byte[] secretKey)
         {
-            DecryptInternal(input, output, null, secretKey.Take(256/8).ToArray());
+            DecryptInternal(input, output, null, secretKey.Take(256 / 8).ToArray());
         }
 
         public static void Decrypt(Stream input, Stream output, string password)
@@ -20,6 +20,8 @@ namespace Encryption
 
         private static void DecryptInternal(Stream input, Stream output, string password, byte[] secretKey)
         {
+            byte[] hmacHash;
+
             // BUG: NO file io!
             var tempPath = Path.GetTempFileName();
             CryptoFileInfo cryptoFileInfo;
@@ -28,25 +30,30 @@ namespace Encryption
                 cryptoFileInfo = CryptoFileInfo.LoadFromDisk(input, rawfile);
             }
 
-            var keyAes = password != null 
-                ? CreateAesKeyFromPassword(password, cryptoFileInfo.Salt, cryptoFileInfo.Iterations) 
+            var keyAes = password != null
+                ? CreateAesKeyFromPassword(password, cryptoFileInfo.Salt, cryptoFileInfo.Iterations)
                 : secretKey;
 
             using (var tempFile = File.OpenRead(tempPath))
             using (var aes = Aes.Create())
             {
-                // BUG: NO Athenticaton
                 aes.Key = keyAes;
                 aes.IV = cryptoFileInfo.Iv;
 
                 using (var encryptor = aes.CreateDecryptor(aes.Key, aes.IV))
                 {
-                    using (var aesStream = new CryptoStream(tempFile, encryptor, CryptoStreamMode.Read))
+                    var hmacsha512 = new HMACSHA512(keyAes);
+                    using (var aesStream = new CryptoStream(output, encryptor, CryptoStreamMode.Write))
+                    using (var hmacStream = new CryptoStream(aesStream, hmacsha512, CryptoStreamMode.Write))
                     {
-                        aesStream.CopyTo(output);
+                        tempFile.CopyTo(hmacStream);
                     }
+                    hmacHash = hmacsha512.Hash;
                 }
             }
+
+            if(!cryptoFileInfo.Hmac.SequenceEqual(hmacHash))
+                throw new CryptographicException("HMAC Hash not as expected");
 
             File.Delete(tempPath);
         }
@@ -63,7 +70,7 @@ namespace Encryption
 
             var iv = RandomHelper.GetRandomData(128);
 
-            EncryptInternal(input, output, keyAes,iv, salt, iterations);
+            EncryptInternal(input, output, keyAes, iv, salt, iterations);
         }
 
         public static void Encrypt(Stream input, Stream output, byte[] secretKey)
@@ -79,21 +86,24 @@ namespace Encryption
         public static void EncryptInternal(Stream input, Stream output, byte[] keyAes, byte[] iv, byte[] salt, int iterations)
         {
             // BUG: NO file io!
+            byte[] hmacHash;
             var tempPath = Path.GetTempFileName();
             using (var tempFile = File.Create(tempPath))
             {
                 using (var aes = Aes.Create())
                 {
-                    // BUG: NO Athenticaton
                     aes.Key = keyAes;
                     aes.IV = iv;
 
                     using (var encryptor = aes.CreateEncryptor(aes.Key, aes.IV))
                     {
-                        using (var aesStream = new CryptoStream(tempFile, encryptor, CryptoStreamMode.Write))
+                        var hmacsha512 = new HMACSHA512(keyAes);
+                        using (var hmacStream = new CryptoStream(tempFile, hmacsha512, CryptoStreamMode.Write))
+                        using (var aesStream = new CryptoStream(hmacStream, encryptor, CryptoStreamMode.Write))
                         {
                             input.CopyTo(aesStream);
                         }
+                        hmacHash = hmacsha512.Hash;
                     }
                 }
             }
@@ -103,6 +113,7 @@ namespace Encryption
                 Iv = iv,
                 Salt = salt,
                 Iterations = iterations,
+                Hmac = hmacHash
             };
 
             using (var tempStream = File.OpenRead(tempPath))
@@ -121,7 +132,5 @@ namespace Encryption
             }
             return keyAes;
         }
-
-
     }
 }
